+++ src/network/devices/linux/command-kernel/createLinuxHostShell.ts
@@ -22,6 +22,7 @@ import { PingCommand, Ping6Command } from './commands/Ping';
 import { TracerouteCommand } from './commands/Traceroute';
 import { ArpingCommand } from './commands/Arping';
 import { UseraddCommand } from './commands/Useradd';
+import { AdduserCommand, AddgroupCommand } from './commands/Adduser';
 import { GroupsCommand } from './commands/Groups';
 import { HeadCommand } from './commands/Head';
 import { IdCommand } from './commands/Id';
@@ -134,6 +135,8 @@ export function createLinuxHostShell(deps: LinuxMachineApiDeps): Interpreter {
   registry.register(() => new TracerouteCommand());
   registry.register(() => new ArpingCommand());
   registry.register(() => new UseraddCommand());
+  registry.register(() => new AdduserCommand());
+  registry.register(() => new AddgroupCommand());
 
   return new Interpreter(registry, machine);
 }